+++ packages/editor-common/web/src/Utils/draftUtils.js
@@ -182,6 +182,45 @@ export const replaceWithEmptyBlock = (editorState, blockKey) => {
   return EditorState.forceSelection(newState, resetBlock.getSelectionAfter());
 };
 
+// export const setSelectionToBlock = (newEditorState, setEditorState, newActiveBlock) => {
+//   const editorState = newEditorState;
+//   const offsetKey = DraftOffsetKey.encode(newActiveBlock.getKey(), 0, 0);
+//   const node = document.querySelectorAll(`[data-offset-key="${offsetKey}"]`)[0];
+//   const selection = window.getSelection();
+//   const range = document.createRange();
+//   range.setStart(node, 0);
+//   range.setEnd(node, 0);
+//   selection.removeAllRanges();
+//   selection.addRange(range);
+
+//   setEditorState(
+//     EditorState.forceSelection(
+//       editorState,
+//       new SelectionState({
+//         anchorKey: newActiveBlock.getKey(),
+//         anchorOffset: 0,
+//         focusKey: newActiveBlock.getKey(),
+//         focusOffset: 0,
+//         isBackward: false,
+//       })
+//     )
+//   );
+// };
+
+// **************************** this function is for oneApp ****************************
+export const createBlockAndFocus = (editorState, data, pluginType) => {
+  const { newBlock, newSelection, newEditorState } = createBlock(editorState, data, pluginType);
+  window.getSelection().removeAllRanges();
+  return new Promise(resolve => {
+    setTimeout(() => {
+      resolve({
+        newEditorState: EditorState.forceSelection(newEditorState, newSelection),
+        newBlock,
+      });
+    }, 0);
+  });
+};
+
 export const createBlock = (editorState, data, type) => {
   const currentEditorState = editorState;
   const contentState = currentEditorState.getCurrentContent();
@@ -192,9 +231,7 @@ export const createBlock = (editorState, data, type) => {
   // when adding atomic block, there is the atomic itself, and then there is a text block with one space,
   // so get the block before the space
   const newBlock = newEditorState.getCurrentContent().getBlockBefore(recentlyCreatedKey);
-
   const newSelection = SelectionState.createEmpty(newBlock.getKey());
-
   return { newBlock, newSelection, newEditorState };
 };